Check that user isn't a site_admin before deleting their blog 1 perms, fixes #536
git-svn-id: http://mu.svn.wordpress.org/trunk@1246 7be80a69-a1ef-0310-a953-fb0f7c49ff36

+++ wp-includes/wpmu-functions.php
@@ -1389,7 +1389,7 @@ function install_blog_defaults($blog_id, $user_id) {
 	$wpdb->query( "DELETE FROM ".$wpdb->usermeta." WHERE  user_id != '".$user_id."' AND meta_key = '".$table_prefix."user_level'" );
 	$wpdb->query( "DELETE FROM ".$wpdb->usermeta." WHERE  user_id != '".$user_id."' AND meta_key = '".$table_prefix."capabilities'" );
 	// Delete any caps that snuck into the previously active blog. (Hardcoded to blog 1 for now.) TODO: Get previous_blog_id.
-	if ( $user_id != 1 )
+	if ( !is_site_admin( $user->user_login ) && $user_id != 1 )
 		$wpdb->query( "DELETE FROM ".$wpdb->usermeta." WHERE  user_id = '".$user_id."' AND meta_key = '" . $wpdb->base_prefix . "1_capabilities'" );
 
 	$wpdb->show_errors();
